move tournaments to announcements panel
