+++ app/public/src/pages/component/announcements/announcements.css
@@ -1,3 +1,12 @@
+.announcements-container {
+    flex: 1;
+    display: flex;
+    flex-flow: column nowrap;
+    justify-content: flex-start;
+    align-items: stretch;
+    overflow-y: auto;
+}
+
 .special-game-announcement {
     background: #40404080;
     padding: 0.4em 0.5em;
@@ -16,4 +25,23 @@
 
 .special-game-announcement span:last-child {
     flex: 0 0 6ch;
+}
+
+.announcements .tournaments {
+    flex: 0;
+    padding: 0;
+
+    & > li {
+        list-style: none;
+        margin: 0.5em 0;
+    }
+}
+
+.announcements .user-chat {
+    flex: 1;
+    min-height: 250px;
+}
+
+.announcements .tournaments li .tournament-item {
+    max-height: 50vh;
 }
