+++ app/public/src/pages/lobby.css
@@ -42,6 +42,7 @@
   font-size: 16px;
   overflow-x: auto;
   overflow-y: hidden;
+  flex: 0 0 auto;
 }
 
 .lobby .current-users-menu {
@@ -57,12 +58,6 @@
   height: 100%;
   display: flex;
   flex-flow: column nowrap;
-  justify-content: flex-start;
-  align-items: stretch;
-}
-
-.lobby .announcements .user-chat {
-  flex: 1;
 }
 
 .lobby h2 {
